django/python-requirements.txt
celery
django
django-model-utils
django-extensions
djangorestframework
markdown
psycopg2-binary
Pillow
pytest-django
redis
requests
uwsgi